@@ -1,8 +1,8 @@
 package com.example.zukkey.arcoresampleforprimer.java;
 
 import android.content.Context;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 import android.widget.Toast;
 
 import com.google.ar.sceneform.utilities.Preconditions;
