@@ -2,10 +2,10 @@ package com.example.zukkey.arcoresampleforprimer.java;
 
 import android.content.Context;
 import android.content.Intent;
-import android.databinding.DataBindingUtil;
+import androidx.databinding.DataBindingUtil;
 import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.v7.app.AppCompatActivity;
+import androidx.annotation.Nullable;
+import androidx.appcompat.app.AppCompatActivity;
 import com.example.zukkey.arcoresampleforprimer.R;
 import com.example.zukkey.arcoresampleforprimer.databinding.ActivityWebBinding;
 
