@@ -2,9 +2,9 @@ package com.example.zukkey.arcoresampleforprimer.misc
 
 import android.app.AlertDialog
 import android.app.Dialog
-import android.databinding.DataBindingUtil
+import androidx.databinding.DataBindingUtil
 import android.os.Bundle
-import android.support.v4.app.DialogFragment
+import androidx.fragment.app.DialogFragment
 import android.view.LayoutInflater
 import com.example.zukkey.arcoresampleforprimer.R
 import com.example.zukkey.arcoresampleforprimer.databinding.EditDialogBinding
